src/modules/opengl_config.py
from OpenGL.GLU import *
from OpenGL.GL import *

# Window dimensions
WINDOW_WIDTH = 800
WINDOW_HEIGHT = 600


def init_gl():
    glClearColor(0.0, 0.0, 0.0, 1.0)
    glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT)
    glEnable(GL_DEPTH_TEST)
    glMatrixMode(GL_PROJECTION)
    glLoadIdentity()
    gluPerspective(45, (WINDOW_WIDTH / WINDOW_HEIGHT), 0.1, 50.0)
    glMatrixMode(GL_MODELVIEW)
