requirements.txt
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==4.4.2
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
imageio==2.29.0
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==9.5.0
proglog==0.1.10
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.4.0
PyOpenGL==3.1.7
pyparsing==3.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.16
